[FLINK-38544][checkpoint] Split spill reader into interface and impl
Make FetchedChannelStateReader an interface so callers depend only on the
contract (nextSegment/snapshot/close/emptyReader); move the implementation
(file stream, two positions, bounded body view) into
FetchedChannelStateReaderImpl. Fold the SpillSegment type into the reader
interface as a nested interface, since a segment only exists relative to
its reader. Callers and tests are untouched except for the nested-type
import; no behavior change.
